refactor(renderer): promote tool-library and publishing out of personal-views/
Step 5 of the renderer reorganization. Tool discovery/installation and
developer-collective publishing are about Tools, not a 'personal view':
- tool-library/             — discovery + installation (was personal-views/tool-library)
- tool-library/publishing/  — developer collectives (was personal-views/tool-publishing)

Pure file moves + import-path updates; no behaviour change.
