+++ src/renderer/src/app/main-dashboard.ts
@@ -43,7 +43,7 @@ import '../groups/elements/entry-title.js';
 import './navigation/groups-sidebar.js';
 import './dialogs/join-group-dialog.js';
 import '../layout/views/cross-group-main.js';
-import '../personal-views/tool-library/tool-library-web2.js';
+import '../tool-library/tool-library-web2.js';
 import '../layout/views/asset-view.js';
 import '../groups/elements/group-container.js';
 import './debugging-panel/debugging-panel.js';
