+++ src/renderer/src/tool-library/elements/curation-list-manager.ts
@@ -2,8 +2,8 @@ import {LitElement, html, css} from "lit";
 import {customElement, state} from "lit/decorators.js";
 import {msg} from "@lit/localize";
 import {ToolCurations} from "@theweave/moss-types";
-import {mossStyles} from "../../../shared-styles";
-import {trashIcon} from "../../../ui/icons";
+import {mossStyles} from "../../shared-styles";
+import {trashIcon} from "../../ui/icons";
 import {DEFAULT_PRODUCTION_TOOL_CURATION_CONFIGS} from "../tool-library-web2";
 
 export interface NamedUrl {
