+++ src/renderer/src/tool-library/elements/install-tool-dialog.ts
@@ -15,11 +15,11 @@
 // import '@shoelace-style/shoelace/dist/components/dialog/dialog.js';
 // import '@shoelace-style/shoelace/dist/components/tooltip/tooltip.js';
 
-// import { groupStoreContext } from '../../../groups/context.js';
-// import { mossStyles } from '../../../shared-styles.js';
-// import { GroupStore } from '../../../groups/group-store.js';
-// import { mossStoreContext } from '../../../context.js';
-// import { MossStore } from '../../../moss-store.js';
+// import { groupStoreContext } from '../../groups/context.js';
+// import { mossStyles } from '../../shared-styles.js';
+// import { GroupStore } from '../../groups/group-store.js';
+// import { mossStoreContext } from '../../context.js';
+// import { MossStore } from '../../moss-store.js';
 // import { Tool, UpdateableEntity } from '@theweave/tool-library-client';
 
 // @localized()
