+++ src/renderer/src/tool-library/elements/tool-bundle-title.ts
@@ -10,9 +10,9 @@
 // import { ActionHash } from '@holochain/client';
 // import { hashProperty } from '@holochain-open-dev/elements';
 
-// import { MossStore } from '../../../moss-store.js';
-// import { mossStoreContext } from '../../../context.js';
-// import { mossStyles } from '../../../shared-styles.js';
+// import { MossStore } from '../../moss-store.js';
+// import { mossStoreContext } from '../../context.js';
+// import { mossStyles } from '../../shared-styles.js';
 // import { Tool, UpdateableEntity } from '@theweave/tool-library-client';
 
 // @customElement('tool-bundle-title')
