+++ src/renderer/src/tool-library/libraryStyles.ts
