+++ src/renderer/src/tool-library/publishing/elements/developer-collective-view.ts
