+++ src/renderer/src/tool-library/publishing/elements/edit-developer-collective.ts
