+++ src/renderer/src/tool-library/publishing/elements/publish-tool.ts
