+++ src/renderer/src/tool-library/publishing/publishing-view.ts
