+++ src/renderer/src/tool-library/tool-library-store.ts
@@ -7,7 +7,7 @@ import {
   retryUntilSuccess,
 } from '@holochain-open-dev/stores';
 import { ActionHash, LazyHoloHashMap, DnaHash, encodeHashToBase64, ProvisionedCell } from '@holochain/client';
-import { ConductorInfo } from '../../electron-api.js';
+import { ConductorInfo } from '../electron-api.js';
 import { Tool, UpdateableEntity, ToolsLibraryClient } from '@theweave/tool-library-client';
 
 export class ToolsLibraryStore {
