+++ src/renderer/src/tool-library/tool-library-web2.ts
@@ -1,7 +1,7 @@
 import { html, LitElement, css } from 'lit';
 import { customElement, query, state } from 'lit/decorators.js';
 import { localized, msg } from '@lit/localize';
-import { getPrimaryVersionBranch } from '../../utils.js';
+import { getPrimaryVersionBranch } from '../utils.js';
 import {
   DeveloperCollective,
   ToolCompatibilityId,
@@ -9,7 +9,7 @@ import {
   ToolCurationList,
   ToolCurator,
 } from '@theweave/moss-types';
-import { UnifiedToolEntry } from '../../types.js';
+import { UnifiedToolEntry } from '../types.js';
 import {
   DEFAULT_PRODUCTION_TOOL_CURATION_CONFIGS,
   fetchUnifiedTools,
@@ -20,31 +20,31 @@ import '@shoelace-style/shoelace/dist/components/icon/icon.js';
 import '@shoelace-style/shoelace/dist/components/button/button.js';
 import '@shoelace-style/shoelace/dist/components/switch/switch.js';
 
-import { mossStyles } from '../../shared-styles.js';
-import '../../groups/elements/invite/select-group.js';
+import { mossStyles } from '../shared-styles.js';
+import '../groups/elements/invite/select-group.js';
 import {mdiEmailOutline, mdiWeb} from '@mdi/js';
 import { wrapPathInSvg } from '@holochain-open-dev/elements';
 import './elements/curation-list-manager.js';
 import './elements/installable-tools-web2.js';
 import './elements/tool-publisher-detail.js';
-import { mossStoreContext } from '../../context.js';
+import { mossStoreContext } from '../context.js';
 import { consume } from '@lit/context';
-import { MossStore } from '../../moss-store.js';
-import { groupStoreContext } from '../../groups/context.js';
-import { GroupStore } from '../../groups/group-store.js';
-import { SelectGroup } from '../../groups/elements/invite/select-group.js';
+import { MossStore } from '../moss-store.js';
+import { groupStoreContext } from '../groups/context.js';
+import { GroupStore } from '../groups/group-store.js';
+import { SelectGroup } from '../groups/elements/invite/select-group.js';
 import { DnaHashB64, decodeHashFromBase64 } from '@holochain/client';
 import { InstallToolDialogWeb2 } from './elements/install-tool-dialog-web2.js';
 import './elements/install-tool-dialog-web2.js';
-import { ToolAndCurationInfo, ToolListUrl } from '../../types';
+import { ToolAndCurationInfo, ToolListUrl } from '../types';
 import {
   appStoreIcon,
   devIcon,
   experimentalToolIcon,
   stableToolIcon,
-} from '../../ui/icons.js';
-import '../../ui/moss-dialog.js';
-import {MossDialog} from "../../ui/moss-dialog";
+} from '../ui/icons.js';
+import '../ui/moss-dialog.js';
+import {MossDialog} from "../ui/moss-dialog";
 import {NamedUrl, UrlListManager} from "./elements/curation-list-manager";
 
 export { DEFAULT_PRODUCTION_TOOL_CURATION_CONFIGS } from './fetch-unified-tools.js';
